:recycle: change package for category recipe search
