+++ backend/src/main/java/net/pengcook/category/service/CategoryService.java
@@ -22,20 +22,9 @@ public class CategoryService {
 
     private final CategoryRepository categoryRepository;
     private final CategoryRecipeRepository categoryRecipeRepository;
-    private final RecipeRepository recipeRepository;
-    private final RecipeService recipeService;
 
     public void saveCategories(Recipe recipe, List<String> categories) {
         categories.forEach(category -> saveCategoryRecipe(recipe, category));
-    }
-
-    public List<MainRecipeResponse> readRecipesOfCategory(RecipeOfCategoryRequest request) {
-        String categoryName = request.category();
-        Pageable pageable = PageRequest.of(request.pageNumber(), request.pageSize());
-        List<Long> recipeIds = categoryRecipeRepository.findRecipeIdsByCategoryName(categoryName, pageable);
-
-        List<RecipeDataResponse> recipeDataResponses = recipeRepository.findRecipeData(recipeIds);
-        return recipeService.convertToMainRecipeResponses(recipeDataResponses);
     }
 
     private void saveCategoryRecipe(Recipe recipe, String name) {
